# backend/v2/openapi.yaml
openapi: 3.0.3
info:
  title: AI Skill Management Platform
  version: 2.0.0
  description: Multi-tenant SaaS for creating, managing, assigning, and executing AI skills
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/skills:
    get:
      summary: List skills (paginated, filtered)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: search, in: query, schema: { type: string, default: "" } }
        - { name: category, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string, default: "" } }
      responses:
        '200':
          description: Paginated skills list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      summary: Create a new skill
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                description: { type: string, maxLength: 5000 }
                content: { type: string, maxLength: 50000 }
                category: { type: string, default: general }
      responses:
        '201':
          description: Skill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/skills/{skill_id}:
    get:
      summary: Get skill by ID
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Full update skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                content: { type: string }
                category: { type: string }
      responses:
        '200':
          description: Skill updated
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      summary: Partial update skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                content: { type: string }
                category: { type: string }
      responses:
        '200':
          description: Skill patched
    delete:
      summary: Soft delete skill (admin)
      tags: [Skills]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill deleted
...
  /api/users:
    get:
      summary: List organization users (paginated)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated users list
...
  /api/org:
    get:
      summary: Get organization profile
      tags: [Organizations]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Organization profile
    patch:
      summary: Update organization settings (admin)
      tags: [Organizations]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                plan: { type: string }
      responses:
        '200':
          description: Organization updated
  /api/org/stats:
    get:
      summary: Get organization dashboard stats
      tags: [Organizations]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Organization statistics
  /api/org/activity:
    get:
      summary: Get organization activity feed (paginated)
      tags: [Organizations]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Activity feed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTHORIZATION_ERROR
              message: Admin role required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Skill not found
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: A skill with this name already exists
  schemas:
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            total: { type: integer }
            per_page: { type: integer }
            total_pages: { type: integer }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: { type: object } }
        meta:
          type: object
          properties:
            page: { type: integer }
            total: { type: integer }
            per_page: { type: integer }
            total_pages: { type: integer }